image processing in sensor.py
